(PE-31696) Require bolt gem before running local tasks
In PE we use the Ruby interpreter shipped with the Puppet Agent to run the
ace server. The ace server is responsible for executing remote tasks.
Remote tasks will often use the puppet library, and run in ACE using the
local transport. Now that puppet has moved to using require_relative
there is a bug where `require 'puppet'` in the task will load Puppet
from the Ruby vendor path instead of the Puppet gem. The vendor Puppet
will then try to autoload gems from the Gem path including Puppet,
causing a `UniqueFile` error when the same file is loaded from different
locations.

This updates ACE to run `-r bolt` before executing tasks on localhost,
which will add Bolt and it's dependencies to the `$LOAD_PATH` including
Puppet, so that the Gem puppet is prefered over the vendor Puppet.